Extract project settings loader into a presenter service
